Fix flaky CI: measure agent latency with perf_counter at sub-ms precision
CI runs the mock pipeline in under 0.05 ms per agent, and latency was
time.time() truncated to int milliseconds - every step rounded to 0 and
test_endpoint_on_live_data failed on p50 > 0 (3.11 job failure on an
otherwise fast runner).

- AgentTrace.duration_ms and PipelineResult.total_latency_ms are floats
  measured with time.perf_counter(), rounded to 3 decimals
- run_metrics and analytics avg latency keep 3-decimal precision
- endpoint test seeds a known run record instead of trusting wall clock
- new regression test pins that 0.012 ms survives rounding

+++ backend/models.py
@@ -176,8 +176,8 @@ class QualityResult(BaseModel):
 class AgentTrace(BaseModel):
     """One hop in the pipeline — kept for auditability and monitoring."""
     agent: str
-    started_at: float = Field(default_factory=time.time)
-    duration_ms: int = 0
+    started_at: float = Field(default_factory=time.time)  # wall clock, for audit
+    duration_ms: float = 0.0  # perf_counter-based; sub-ms precision so fast runs never truncate to 0
     input_summary: str = ""
     output_summary: str = ""
     tokens_in: int = 0
@@ -200,7 +200,7 @@ class PipelineResult(BaseModel):
     review_id: str | None = None
     trace: list[AgentTrace] = Field(default_factory=list)
     total_cost_usd: float = 0.0
-    total_latency_ms: int = 0
+    total_latency_ms: float = 0.0
     error: str | None = None
